feat: Add task creation functionality to TaskManager
This commit adds the ability to create tasks in the TaskManager class. The `add_task` method now checks if a user is logged in before allowing the task to be added. If no user is logged in, an error message is printed. Otherwise, a new Task object is created with the provided title, description, and the username of the logged-in user. The task is then appended to the list of tasks in the TaskManager. A success message is printed to indicate that the task has been successfully added.